handle multiple playback delete requests

+++ swfologger.py
@@ -24,7 +24,10 @@ while True:
                 os.rename("swfolog.txt","swfoplayback.txt")
             os.system("python3 swfoplayback.py &")
         elif data[0] == 0x44 and data[1] == 0x50:  # DP
-            print("Deleted Playback File\r\n")
-            os.remove("swfoplayback.txt")
+            if os.path.exists("swfoplayback.txt"):
+                print("Deleted Playback File\r\n")
+                os.remove("swfoplayback.txt")
+            else:
+                print("No Playback File found\r\n")
     else:
         print(".",end = '', flush = True)
